+++ src/mem.h
@@ -6,10 +6,10 @@
 class Memory
 {
   public:
-    Memory(unsigned int size);
+    Memory(uint16_t size);
 
-    uint8_t read(unsigned int address);
-    void write(unsigned int address, uint8_t b);
+    uint8_t read(uint16_t address);
+    void write(uint16_t address, uint8_t b);
 
   private:
     unsigned int size;
